feat(model): add full Mamba2 model, SSM inference cache, detection, and registry
Restructure model implementations into proper module hierarchies and add
new high-level model infrastructure:

Mamba2:
- Split mamba2.rs into mamba2/{config,layer} and ssm/{types,scan,inference}
- Add ssm/inference.rs with stateful step functions for autoregressive
  decoding (ssm_forward_sequential_inference, ssm_step_inference)
- Add Mamba2Model: full embedding + N layers + norm + LM head

LLaMA:
- Split llama/model.rs into llama/model/{blocks,forward} for size compliance

Architecture detection:
- Add model/detection/ with detectors for GGUF, SafeTensors, and
  HuggingFace config.json formats
- Expose detect_architecture and detect_architecture_from_names

Model registry:
- Add model/registry.rs with LoadedModel<R> enum dispatching over
  concrete model types with a unified inference API

+++ src/model/llama/model_tp.rs
@@ -129,7 +129,7 @@ impl<R: Runtime<DType = DType>> LlamaTp<R> {
             let block = LlamaBlockTp {
                 input_layernorm: RmsNorm::new(
                     Tensor::<R>::ones(&[hidden], dt, device),
-                    config.rms_norm_eps,
+                    config.rms_norm_eps as f32,
                     true,
                 ),
                 self_attn: LlamaAttentionTp {
@@ -160,7 +160,7 @@ impl<R: Runtime<DType = DType>> LlamaTp<R> {
                 },
                 post_attention_layernorm: RmsNorm::new(
                     Tensor::<R>::ones(&[hidden], dt, device),
-                    config.rms_norm_eps,
+                    config.rms_norm_eps as f32,
                     true,
                 ),
                 mlp: LlamaMlpTp {
@@ -188,7 +188,7 @@ impl<R: Runtime<DType = DType>> LlamaTp<R> {
         // Final norm (replicated)
         let norm = RmsNorm::new(
             Tensor::<R>::ones(&[hidden], dt, device),
-            config.rms_norm_eps,
+            config.rms_norm_eps as f32,
             true,
         );
 
@@ -276,7 +276,7 @@ impl<R: Runtime<DType = DType>> LlamaTp<R> {
             let block = LlamaBlockTp {
                 input_layernorm: RmsNorm::new(
                     layer_vb.take_tensor("input_layernorm.weight")?,
-                    config.rms_norm_eps,
+                    config.rms_norm_eps as f32,
                     false,
                 ),
                 self_attn: LlamaAttentionTp {
@@ -290,7 +290,7 @@ impl<R: Runtime<DType = DType>> LlamaTp<R> {
                 },
                 post_attention_layernorm: RmsNorm::new(
                     layer_vb.take_tensor("post_attention_layernorm.weight")?,
-                    config.rms_norm_eps,
+                    config.rms_norm_eps as f32,
                     false,
                 ),
                 mlp: LlamaMlpTp {
@@ -304,7 +304,7 @@ impl<R: Runtime<DType = DType>> LlamaTp<R> {
 
         let norm = RmsNorm::new(
             model_vb.take_tensor("norm.weight")?,
-            config.rms_norm_eps,
+            config.rms_norm_eps as f32,
             false,
         );
